Add object for representing additional token info TokenInfoObject #4695
